+++ subsys/settings/src/CMakeLists.txt
@@ -11,3 +11,4 @@ zephyr_sources_ifdef(CONFIG_SETTINGS_RUNTIME settings_runtime.c)
 zephyr_sources_ifdef(CONFIG_SETTINGS_FS settings_file.c)
 zephyr_sources_ifdef(CONFIG_SETTINGS_FCB settings_fcb.c)
 zephyr_sources_ifdef(CONFIG_SETTINGS_NVS settings_nvs.c)
+zephyr_sources_ifdef(CONFIG_SETTINGS_NONE settings_none.c)
