+++ subsys/settings/src/settings_none.c
@@ -1,0 +1,10 @@
+/*
+ * Copyright (c) 2019 Nordic Semiconductor ASA
+ *
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+int settings_backend_init(void)
+{
+	return 0;
+}
